eutonafila/requirements.txt
# Core
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
uvicorn==0.25.0
gunicorn==21.2.0

# Async and real-time
channels==4.0.0
daphne==4.0.0
celery==5.5.1
django-celery-results==2.6.0
pika==1.3.2

# Optional: Choose one for production
# redis==5.0.1
# kombu==5.5.3

# WebSocket connections
channels_redis==4.1.0

# Development
black==24.1.1
isort==5.13.2
pylint==3.0.3
pytest==7.4.3
pytest-django==4.7.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0 